Refactor out UsfmProjectReader.
